# using a table-based convert_hex_to_binary function for better speed